changed the signature of methods; changed names
