+++ app/src/main/java/com/example/annakocheshkova/testapplication/Receiver/MyBootReceiver.java
@@ -3,7 +3,6 @@ package com.example.annakocheshkova.testapplication.Receiver;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import com.example.annakocheshkova.testapplication.Utils.NotificationAlarmManager;
 
 /**
  * Listener receiving boot complete event, reschedules all the notifications after reboot
@@ -17,6 +16,6 @@ public class MyBootReceiver extends BroadcastReceiver
      */
     @Override
     public void onReceive(Context context, Intent intent) {
-        NotificationAlarmManager.scheduleAlarms();
+        ReminderAlarmManager.scheduleAlarms();
     }
 }
